Refactor RSpec config
