+++ .ruby-version
@@ -1,1 +1,1 @@
-ruby-2.6.5
+2.6.5
