--- .travis.yml
@@ -1,9 +1,0 @@
-language: ruby
-rvm:
- - 2.6.5
-services:
-  - postgresql
-before_script:
-  - cp config/database.yml.travis config/database.yml
-  - RAILS_ENV=test bundle exec rake db:create
-script: 'bundle exec rspec'
