+++ spec/support/config/rspec_file_chef.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+RSpec.configure do |config|
+  RspecFileChef::FileChef.configure do |configuration|
+    configuration.rspec_path = File.expand_path(__dir__)
+  end
+end
